fix(events): pass through updated model values
The current behaviour is that event handlers will be
passed the value of each argument as it was at the time
the event handler was created, which is when the
directive is linked. This means an event handler being
passed args derived from the scope will not receive
any updated values.

This commit will alter the behaviour to send through
any updated model values.